fix: wrapscan tests
