+++ packages/config-bundles/polywrap-sys-config-bundle/polywrap_sys_config_bundle/embeds/ipfs-sync-resolver/wrap.wasm
