+++ packages/config-bundles/polywrap-sys-config-bundle/tests/test_sanity.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from polywrap_client_config_builder import PolywrapClientConfigBuilder
-from polywrap_core import Uri, UriResolutionContext
+from polywrap_core import Uri, UriPackage
 from polywrap_client import PolywrapClient
 from polywrap_sys_config_bundle import get_sys_config
 
@@ -33,22 +33,48 @@ def test_file_system_resolver():
 
     assert response["manifest"]
 
+    uri_package = client.try_resolve_uri(
+        uri=Uri.from_str(f"wrap://fs/{path_to_resolve}")
+    )
+    assert uri_package
+    assert isinstance(uri_package, UriPackage)
+
 
 def test_http_resolver():
     config = PolywrapClientConfigBuilder().add(get_sys_config()).build()
     client = PolywrapClient(config)
+    http_path = "wraps.wrapscan.io/r/polywrap/wrapscan-uri-resolver@1.0"
 
     response = client.invoke(
         uri=Uri.from_str("ens/wraps.eth:http-uri-resolver-ext@1.0.1"),
         method="tryResolveUri",
-        args={"authority": "https", "path": "wraps.wrapscan.io/r/polywrap/wrapscan-uri-resolver@1.0"},
+        args={"authority": "https", "path": http_path},
     )
 
     assert response["uri"]
     assert Uri.from_str(response["uri"]).authority == "ipfs"
 
+    uri_package = client.try_resolve_uri(
+        uri=Uri.from_str(f"wrap://https/{http_path}")
+    )
+    assert uri_package
+    assert isinstance(uri_package, UriPackage)
 
-def test_resolve_wrapscan_resolver():
+
+
+def test_ipfs_resolver():
+    config = PolywrapClientConfigBuilder().add(get_sys_config()).build()
+    client = PolywrapClient(config)
+
+    result = client.try_resolve_uri(
+        uri=Uri.from_str("wrap://ipfs/QmfRCVA1MSAjUbrXXjya4xA9QHkbWeiKRsT7Um1cvrR7FY")
+    )
+
+    assert result is not None
+    assert isinstance(result, UriPackage)
+
+
+def test_can_resolve_wrapscan_resolver():
     config = PolywrapClientConfigBuilder().add(get_sys_config()).build()
     client = PolywrapClient(config)
     response = client.try_resolve_uri(
@@ -56,5 +82,15 @@ def test_resolve_wrapscan_resolver():
     )
 
     assert response
-    assert isinstance(response, Uri)
-    assert response.authority == "ipfs"
+    assert isinstance(response, UriPackage)
+
+
+def test_wrapscan_resolver():
+    config = PolywrapClientConfigBuilder().add(get_sys_config()).build()
+    client = PolywrapClient(config)
+    response = client.try_resolve_uri(
+        Uri("wrapscan.io", "polywrap/uri-resolver@1.0"),
+    )
+
+    assert response
+    assert isinstance(response, UriPackage)
