--- packages/config-bundles/polywrap-web3-config-bundle/polywrap_web3_config_bundle/embeds/__init__.py
@@ -1,16 +1,0 @@
-"""This module contains the embedded wraps for the system configuration bundle."""
-
-from pathlib import Path
-
-from polywrap_core import WrapPackage
-from polywrap_sys_config_bundle.types import EmbeddedFileReader
-from polywrap_wasm import WasmPackage
-
-
-def get_embedded_wrap(name: str) -> WrapPackage:
-    """Get the embedded wrap with the given name."""
-    embedded_wrap_path = Path(__file__).parent / name
-    return WasmPackage(EmbeddedFileReader(embedded_wrap_path))
-
-
-__all__ = ["get_embedded_wrap"]
